+++ craft_application/services/lifecycle.py
@@ -148,6 +148,7 @@ class LifecycleService(base.AppService):
         self._project = self._services.get("project").get()
         self._lcm = self._init_lifecycle_manager()
         callbacks.register_post_step(self.post_prime, step_list=[Step.PRIME])
+        callbacks.register_configure_overlay(repositories.enable_overlay_eol)
 
     def _get_build(self) -> craft_platforms.BuildInfo:
         """Get the build for this run."""